+++ src/main/java/com/example/demo/farmerEntity.java
@@ -10,7 +10,7 @@ public class farmerEntity {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
-	private int id;
+	private long id;
 	
 	private String name;
 	
@@ -43,11 +43,13 @@ public class farmerEntity {
 
 	
 	
-	public int getId() {
+	
+
+	public long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
